metaboxes fixed

+++ src/lib/php-compat/WpOptions.php
@@ -162,7 +162,7 @@ class WpOptions
     {
         return $this->hasMetaBoxData;
     }
-    
+        
     /**
      * Agrega la página de opciones en el administrador y la funcion del metabox si es necesaria
      * @access public
@@ -171,21 +171,21 @@ class WpOptions
     {
         if(function_exists('add_object_page'))
         {
-            add_object_page(_s('Configure ') . $this->themeName, $this->themeName, 8, basename(__FILE__),  $this->getFunctionScope('render'),  $this->menuIcon);
+            add_object_page(_s('Configure ') . $this->themeName, $this->themeName, 'edit_themes', basename(__FILE__),  $this->getFunctionScope('render'),  $this->menuIcon);
         }
         else
         {
-            add_menu_page(_s('Configure ') . $this->themeName, $this->themeName, 8, basename(__FILE__),  $this->getFunctionScope('render'),  $this->menuIcon);
+            add_menu_page(_s('Configure ') . $this->themeName, $this->themeName, 'edit_themes', basename(__FILE__),  $this->getFunctionScope('render'),  $this->menuIcon);
         }
         
         foreach($this->subpages as $sub)
         {
-            add_submenu_page(basename(__FILE__), _s($sub['pageTitle']), _s($sub['title']), 8, $sub['slug'], $sub['function']);
+            add_submenu_page(basename(__FILE__), _s($sub['pageTitle']), _s($sub['title']), 'edit_themes', $sub['slug'], $sub['function']);
         }
         
         if ($this->hasMetaBox())
         {
-            add_meta_box('wpoptions_section', $this->themeName . ' :: '._s("Post Settings").'', $this->getFunctionScope('renderMetaBox'), 'post', 'normal');
+            add_meta_box('wpoptions_section', $this->themeName . ' :: '._s("Post Settings"), $this->getFunctionScope('renderMetaBox'), 'post', 'advanced');
             add_action('save_post', $this->getFunctionScope('savePostData'));
         }
     }
@@ -768,7 +768,7 @@ class WpOptions
      */
     function savePostData($idPost)
     {
-        if ( !wp_verify_nonce( $_POST['wpoptions_nonce'], plugin_basename(__FILE__) ))
+        if (isset($_POST['wpoptions_nonce']) && !wp_verify_nonce( $_POST['wpoptions_nonce'], plugin_basename(__FILE__) ))
             return $post_id;
 
         if ( defined('DOING_AUTOSAVE') && DOING_AUTOSAVE ) 
@@ -782,22 +782,26 @@ class WpOptions
         
         foreach($this->optionsInMetaBox as $option)
         {
-            $option->setDbSource($option->Sources['POST_META']);
-            $option->setInputName($this->getCamelCase('wp_options') . '_' . $this->baseThemeName);
-            $option->setDefaultValue('');
-            $option->setValue('');
-            if(is_array($_POST[$option->getFormName()]))
-                $data = $_POST[$option->getFormName()];
-            else
-                $data = (get_magic_quotes_gpc()) ? stripslashes($_POST[$option->getFormName()]) : $_POST[$option->getFormName()]; 
-            $data = $option->set($data);
-            
-            if(get_post_meta($idPost, $option->getName() . '_value') == "")
-                add_post_meta($idPost, $option->getName() . '_value', $data, true);
-            elseif($data != get_post_meta($idPost, $option->getName() . '_value', true))
-                update_post_meta($idPost, $option->getName() . '_value', $data);
-            elseif($data == "")
-                delete_post_meta($idPost, $option->getName() . '_value', get_post_meta($idPost, $option->getName() . '_value', true));
+            $option->setDbSource(WpOption::$Sources['POST_META']);
+            if(isset($_POST[$option->getFormName()]))
+            {
+                $option->setDbSource($option->Sources['POST_META']);
+                $option->setInputName($this->getCamelCase('wp_options') . '_' . $this->baseThemeName);
+                $option->setDefaultValue('');
+                $option->setValue('');
+                if(is_array($_POST[$option->getFormName()]))
+                    $data = $_POST[$option->getFormName()];
+                else
+                    $data = (get_magic_quotes_gpc()) ? stripslashes($_POST[$option->getFormName()]) : $_POST[$option->getFormName()]; 
+                $data = $option->set($data);
+                
+                if(get_post_meta($idPost, $option->getName() . '_value') == "")
+                    add_post_meta($idPost, $option->getName() . '_value', $data, true);
+                elseif($data != get_post_meta($idPost, $option->getName() . '_value', true))
+                    update_post_meta($idPost, $option->getName() . '_value', $data);
+                elseif($data == "")
+                    delete_post_meta($idPost, $option->getName() . '_value', get_post_meta($idPost, $option->getName() . '_value', true));
+            }
         }
     }
     
@@ -1191,7 +1195,7 @@ class WpOptions
                     <p class='submit'><input type='submit' class='button-primary' value='"._s('Save changes')."' />
                 
                 <input type='hidden' name='action' id='action' value='update-wp-options' />
-                ".wp_nonce_field('update-wp-options')."
+                ".wp_nonce_field('update-wp-options','_wpnonce',true,false)."
                 </form>
                 <h2>"._s('Delete Theme options')."</h2>
                 <p>"._s('To completely remove these theme options from your database (reminder: they are all stored in Wordpress options table')." <em>{$this->wpdb->options}</em>),
